% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2_dataGenerating.R
\name{align_scans}
\alias{align_scans}
\title{Get scan time and align between ms events}
\usage{
align_scans(msexp, rt_range, sample_idx)
}
\arguments{
\item{msexp}{\code{MsExperiment} object obtained from xcms or with \code{DIANMF::detect_xcms_peaks()}.}

\item{rt_range}{retention time range}

\item{sample_idx}{subset a sample}
}
\value{
\code{list} of \code{data.table} \code{data.frame} objects: raw_dt and time_dic.
}
\description{
Get scan time and align between ms events
}
