% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2_dataGenerating.R
\name{build_mixed_matrix}
\alias{build_mixed_matrix}
\title{Create the mixed matrix}
\usage{
build_mixed_matrix(xicdt, nscans = 4)
}
\arguments{
\item{xicdt}{\code{data.table} of the xics build with \code{DIANMF::build_XICs()}}

\item{nscans}{number of minimal consecutive value to check}
}
\value{
\code{list} of the mixed matrices (ions x scans) with two levels:by
- [1] mixed matrix with more than nscans consectuvies scans
- [2] mixed matrix with less than nscans consectuvies scans
}
\description{
Create the mixed matrix
}
