% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2_dataGenerating.R
\name{build_ms2XICs}
\alias{build_ms2XICs}
\title{Build MS2 xics}
\usage{
build_ms2XICs(msexp, raw_dt, time_dic, rt_range, MS2_ISOEACHL = T)
}
\arguments{
\item{msexp}{\code{MsExperiment} object obtained from xcms or with \code{DIANMF::detect_xcms_peaks()}.}

\item{raw_dt}{\code{data.frame} from \code{DIANMF::get_rawD_ntime}.}

\item{time_dic}{\code{data.frame} from \code{DIANMF::get_rawD_ntime}.}

\item{rt_range}{retention time range.}

\item{MS2_ISOEACHL}{\code{logical} is TRUE to build the MS2 xics from raw data, else from xcms peaks.}
}
\value{
\code{data.table} \code{data.frame} MS2 EICs.
}
\description{
Build MS2 xics
}
