% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2_dataGenerating.R
\name{build_XICs}
\alias{build_XICs}
\title{Build MS1 xics from peak list}
\usage{
build_XICs(peaks_dt, rawdt, method = c("all", "max")[2])
}
\arguments{
\item{peaks_dt}{\code{data.table} \code{data.frame} targeted peaks in one rt_range.}

\item{rawdt}{\code{data.frame} from \code{DIANMF::get_rawD_ntime}}

\item{method}{\code{string} to select a method for returning data:
\itemize{
\item \code{all}: return all points in range.
\item \code{max}: return max point by retention time.
}}
}
\value{
\code{data.table} \code{data.frame} MS1 EICs.
}
\description{
Build MS1 xics from peak list
}
