% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modified_Decometdia.R
\name{CatLineSeparator}
\alias{CatLineSeparator}
\title{separate info by line}
\usage{
CatLineSeparator(info)
}
\arguments{
\item{info}{character info to be printed}
}
\value{
print info separeted by line
}
\description{
separate info by line
}
