% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank.R
\name{concordance_rank}
\alias{concordance_rank}
\title{determine the rank of factorization using concordance metric}
\usage{
concordance_rank(mat.m)
}
\arguments{
\item{mat.m}{matrix}
}
\value{
rank of mat.m
}
\description{
determine the rank of factorization using concordance metric
}
