% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_peaks_for_precursors_functions.R
\name{create_list_function.l}
\alias{create_list_function.l}
\title{create list for the spiked precursors/compounds}
\usage{
create_list_function.l(compound.df)
}
\arguments{
\item{compound.df}{table, data-frame of the spiked compounds}
}
\value{
empty list of all precursors I'm searching for
}
\description{
create list for the spiked precursors/compounds
}
