% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step1_dataPreparation.R
\name{create_seq}
\alias{create_seq}
\title{Create simple sequence from a list of mzML paths}
\usage{
create_seq(mzML_path)
}
\arguments{
\item{mzML_path}{A character vector with mzML paths.}
}
\value{
a data.table corresponding to a sequence.
}
\description{
Create simple sequence from a list of mzML paths
}
