% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_example.R
\docType{data}
\name{data_example}
\alias{data_example}
\title{This is data to be included in my package}
\description{
This is data to be included in my package
}
\author{
Diana Karaki \email{diana.karaki@cea.fr}
}
\keyword{data}
