% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modified_Decometdia.R
\name{DecoMetDIA_modified_funct}
\alias{DecoMetDIA_modified_funct}
\title{get dia_features and process the data using xcms}
\usage{
DecoMetDIA_modified_funct(
  d.in,
  d.out,
  nSlaves = 6,
  peakwidth = c(5, 30),
  mzdiff = 0.001,
  sn = 0,
  ppm.pd = 6,
  polarity = "positive",
  lc = "RPLC",
  minfrac = 0.5,
  bw = 5,
  mzwid = 0.015,
  is.ms1.filter = FALSE,
  thr.ms1.filter = NULL,
  filed.ms1.filter = c("maxo", "into", "intb"),
  is.plot.eic.feature = F,
  is.check.eic = FALSE,
  is.PlotEIC.MS2 = FALSE,
  snthr.ms2 = 5,
  fp.swath.setup,
  ppm.ms2.mtp = 15,
  int.filter = 50,
  minfrac.vote = 0.5,
  is.plot.eic.spec = F,
  is.output.spec.smp = TRUE,
  is.span.constant = FALSE,
  rerun = FALSE,
  rerun.ms2 = FALSE,
  is.ms1.only = FALSE,
  is.dec.all = FALSE,
  is.dec.smoothed = FALSE,
  is.norm.consensus = TRUE,
  correlation = TRUE
)
}
\arguments{
\item{d.in}{character}

\item{d.out}{path results path
MS1 peaks detection parameters}

\item{nSlaves}{integer}

\item{peakwidth}{integer range of the chromatogram peak width in seconds}

\item{mzdiff}{integer}

\item{sn}{integer singl to noise ration}

\item{ppm.pd}{integer max mz devaition in consecutive scans}

\item{polarity}{string 'POS' or 'NEG'}

\item{lc}{string Liquid chromatoghraphy column used for chromatograms separation}

\item{minfrac}{integer}

\item{bw}{integer}

\item{mzwid}{integer}

\item{is.ms1.filter}{Boolean}

\item{thr.ms1.filter}{Boolean or NULL}

\item{filed.ms1.filter}{character}

\item{is.plot.eic.feature}{Boolean}

\item{is.check.eic}{Boolean
MS2 peaks detections and processed using the SWATH (dictionary) file}

\item{is.PlotEIC.MS2}{Boolean}

\item{snthr.ms2}{integer}

\item{fp.swath.setup}{SWATH file path}

\item{ppm.ms2.mtp}{integer}

\item{int.filter}{integer}

\item{minfrac.vote}{integer}

\item{is.plot.eic.spec}{Boolean}

\item{is.output.spec.smp}{Boolean}

\item{is.span.constant}{Boolean}

\item{rerun}{Boolean}

\item{rerun.ms2}{Boolean}

\item{is.ms1.only}{Boolean}

\item{is.dec.all}{Boolean}

\item{is.dec.smoothed}{Boolean}

\item{is.norm.consensus}{Boolean
Parameter I added it, if you have several files, you can choose to process them in correlation or every file separately}

\item{correlation}{Boolean}
}
\value{
dia.features and xcms objects
}
\description{
get dia_features and process the data using xcms
}
