% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step1_detect_peaks.R
\name{detect_EICs.f}
\alias{detect_EICs.f}
\title{Detect EICs peaks}
\usage{
detect_EICs.f(
  rawData.onDiskMSnExp,
  d.out = NULL,
  ppm = 6,
  peakwidth = c(6, 60),
  snthresh = 1,
  prefilter = c(5, 4000),
  mzCenterFun = "wMeanApex3",
  integrate = 2,
  mzdiff = -0.001,
  noise = 0,
  firstBaselineCheck = FALSE
)
}
\arguments{
\item{rawData.onDiskMSnExp}{}

\item{d.out}{path results path}

\item{ppm}{integer max mz devaition in consecutive scans}

\item{peakwidth}{integer range of the chromatogram peak width in seconds}

\item{snthresh}{integer singl to noise ration}

\item{prefilter}{integer}

\item{mzCenterFun}{string method}

\item{integrate}{integer}

\item{mzdiff}{integer}

\item{noise}{}

\item{firstBaselineCheck}{}
}
\value{
rawData.XCMSnExp
}
\description{
Detect EICs peaks
}
