% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step1_dataPreparation.R
\name{detect_xcms_peaks}
\alias{detect_xcms_peaks}
\title{Detect MS1 peaks using XCMS.}
\usage{
detect_xcms_peaks(
  sequence_table,
  params = list(CentWaveParam = xcms::CentWaveParam(), MergeNeighboringPeaksParam =
    xcms::MergeNeighboringPeaksParam(), ObiwarpParam = xcms::ObiwarpParam(),
    PeakDensityParam = xcms::PeakDensityParam(sampleGroups = NA), ChromPeakAreaParam =
    xcms::ChromPeakAreaParam()),
  rt_range = NULL
)
}
\arguments{
\item{sequence_table}{A data.frame with at least (mzml_path, class, InjectionOrder).}

\item{params}{A list to set the parameters from xcms pipeline steps.
List names should be xcms parameters methods (ex: \verb{list("CentWaveParam" = CentWaveParam()})}

\item{rt_range}{(Optional) rt range to subset before running peak detection}
}
\value{
MsExperiment objects
}
\description{
Detect MS1 peaks using XCMS.
}
