% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_dia_features.R
\name{dia_features.f}
\alias{dia_features.f}
\title{get raw dia features from a mzML file}
\usage{
dia_features.f(
  files,
  d.out,
  ppm.pd = 6,
  sn = 0,
  peakwidth = c(5, 30),
  mzdiff = 3e-04,
  nSlaves
)
}
\arguments{
\item{files}{list_of_strings mzML files paths}

\item{d.out}{path results path}

\item{ppm.pd}{integer max mz devaition in consecutive scans}

\item{sn}{integer singl to noise ration}

\item{peakwidth}{integer range of the chromatogram peak width in seconds}

\item{mzdiff}{integer}

\item{nSlaves}{integer}
}
\value{
dia.feature extracted using xcms
}
\description{
get raw dia features from a mzML file
}
