% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step4_processing.R
\name{DIANMF.f}
\alias{DIANMF.f}
\title{Process xcms object}
\usage{
DIANMF.f(
  msexp,
  d.out = FALSE,
  sample_idx = 1,
  MS2_ISOEACHL = T,
  MS1MS2_L = F,
  rank = 10,
  maximumIteration = 200,
  maxFBIteration = 100,
  toleranceFB = 1e-05,
  initialization_method = "nndsvd",
  errors_print = FALSE,
  method = "svsd",
  scan_rt_ext = 10,
  min_distance = 5
)
}
\arguments{
\item{msexp}{\code{MsExperiment} object obtained from xcms or with \code{DIANMF::detect_xcms_peaks()}.}

\item{d.out}{\code{logical} TRUE to save results, else FALSE}

\item{sample_idx}{NULL to process all files else, \code{numeric} index of the sample.}

\item{MS2_ISOEACHL}{\code{logical}}

\item{MS1MS2_L}{\code{logical}}

\item{rank}{\code{numeric(1)} number of pure compounds in the mixed matrix; rank of factorization.}

\item{maximumIteration}{\code{numeric(1)} maximum number of iterations.}

\item{maxFBIteration}{\code{numeric(1)} maximum number of iterations in the FB algorithm.}

\item{toleranceFB}{\code{numeric(1)} tolerance or stopping value of the FB algorithm.}

\item{initialization_method}{\code{character} to specify the initialization method: random, nndsvd or subSample.}

\item{errors_print}{\code{Logical} \code{TRUE} to print the error difference, \code{FALSE} otherwise.}

\item{method}{Method to compute svd: "base", "fast" or "svds"}

\item{scan_rt_ext}{\code{numeric}}

\item{min_distance}{\code{numeric}}
}
\value{
\code{list}
}
\description{
Process xcms object
}
