% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_algo.R
\name{error_function}
\alias{error_function}
\title{calculate the difference between two matrices}
\usage{
error_function(Y, A, S)
}
\arguments{
\item{Y}{matrix}

\item{A}{matrix elution profiles}

\item{S}{matrix pure spectra}
}
\value{
numeric the error between Y and its approximation AS
}
\description{
calculate the difference between two matrices
}
