% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exporting_fcts.R
\name{export_featureSpect}
\alias{export_featureSpect}
\title{Extract \code{Spectra} object of a feature}
\usage{
export_featureSpect(
  features.l,
  feature_id,
  sample_index,
  type = c("pure", "mixed")[1],
  method = c("all", "best")[2],
  max_method = c("contribution", "max_value")[2]
)
}
\arguments{
\item{features.l}{\code{list} obtained with \code{DIANMF::DIANMF.f()}}

\item{feature_id}{\code{character} feature id.}

\item{sample_index}{\code{numeric(1)} sample index.}

\item{type}{\code{string} to retrieve the "pure" or "mixed" elution profiles.}

\item{method}{\code{string} to extract "all" sources or just the "best" source.}

\item{max_method}{\code{string} to choose a method to select the optimal iteration:
\itemize{
\item \code{contribution}: select the first iteration with maximum contribution.
\item \code{max_value}: select the first iteration with maximum apex pure value.
}}
}
\value{
\code{Spectra} object of a feature
}
\description{
Extract \code{Spectra} object of a feature
}
