% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exporting_fcts.R
\name{exportMSFeaturesSpectra}
\alias{exportMSFeaturesSpectra}
\title{Export Features MS spectra}
\usage{
exportMSFeaturesSpectra(ms1features.df, features.l)
}
\arguments{
\item{ms1features.df}{\code{data.frame} ms1 features obtained by xcms or with \code{DIANMF::extract_xcms_features}.}

\item{features.l}{\code{list} created by \code{DIANMF::DIANM.f}}
}
\value{
\code{list} of \code{Spectra} objects for every feature.
}
\description{
Export Features MS spectra
}
