% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exporting_fcts.R
\name{exportMSSpectra}
\alias{exportMSSpectra}
\title{Title}
\usage{
exportMSSpectra(features.l)
}
\arguments{
\item{features.l}{\code{list}}
}
\value{
\code{list} of \code{Spectra} objects
}
\description{
Title
}
