% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_dia_features.R
\name{extract_dia_features.f}
\alias{extract_dia_features.f}
\title{get raw dia features from a mzML file}
\usage{
extract_dia_features.f(
  d.in,
  d.out,
  nSlaves = 6,
  peakwidth = c(5, 30),
  mzdiff = 0.001,
  sn = 0,
  ppm.pd = 6,
  rerun = FALSE,
  correlation = FALSE
)
}
\arguments{
\item{d.in}{mzML files paths}

\item{d.out}{path results path}

\item{nSlaves}{integer}

\item{peakwidth}{integer range of the chromatogram peak width in seconds}

\item{mzdiff}{integer}

\item{sn}{integer singl to noise ration}

\item{ppm.pd}{integer max mz devaition in consecutive scans}

\item{rerun}{boolean}

\item{correlation}{boolean}
}
\value{
extract all dia features of all mzML files
}
\description{
get raw dia features from a mzML file
}
