% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_parameters.R
\name{extract_ms1_ms2_data_parameters}
\alias{extract_ms1_ms2_data_parameters}
\title{specify the parameters for MS1 & MS2 data extraction for every MS1 peak}
\usage{
extract_ms1_ms2_data_parameters(ppm.ms2.mtp = 15, int.filter = 50)
}
\arguments{
\item{ppm.ms2.mtp}{integer}

\item{int.filter}{integer}
}
\value{
list of parameters to extract mixed matrices and there information
}
\description{
specify the parameters for MS1 & MS2 data extraction for every MS1 peak
}
