% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step3_create_mixed_mat.R
\name{extract_ms2_matrix.f}
\alias{extract_ms2_matrix.f}
\title{Extract MS2 data}
\usage{
extract_ms2_matrix.f(
  idx.pg,
  eics_peaks.mat,
  rawData.onDiskMSnExp,
  ppm = 7,
  rt_index = TRUE,
  mz_range = TRUE
)
}
\arguments{
\item{idx.pg}{}

\item{eics_peaks.mat}{}

\item{rawData.onDiskMSnExp}{}

\item{ppm}{}

\item{rt_index}{}

\item{mz_range}{}
}
\value{
mixed MS2 matrix
}
\description{
Extract MS2 data
}
