% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step1_dataPreparation.R
\name{extract_xcms_features}
\alias{extract_xcms_features}
\title{Extract MS1 features}
\usage{
extract_xcms_features(msexp, orderL = F, orderL_sample = NULL)
}
\arguments{
\item{msexp}{\code{MsExperiment} object obtained from xcms or with \code{DIANMF::detect_xcms_peaks()}.}

\item{orderL}{\code{logical} TRUE to order the features by intensity order of orderL_sample.}

\item{orderL_sample}{\code{character} sample name.}
}
\value{
\code{data.table} \code{data.frame} object.
}
\description{
Extract MS1 features
}
