% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step1_dataPreparation.R
\name{extract_xcms_peaks}
\alias{extract_xcms_peaks}
\title{Extract MS1 peaks}
\usage{
extract_xcms_peaks(msexp)
}
\arguments{
\item{msexp}{\code{MsExperiment} object obtained from xcms or with \code{DIANMF::detect_xcms_peaks()}.}
}
\value{
MS1 peaks \code{matrix}.
}
\description{
Extract MS1 peaks
}
