% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_parameters.R
\name{factorization_parameters}
\alias{factorization_parameters}
\title{specify the factorization algorithm parameters}
\usage{
factorization_parameters(
  maximumIteration = 50,
  maxFBIteration = 10,
  toleranceFB = 1e-04,
  useTranspose = TRUE
)
}
\arguments{
\item{maximumIteration}{integer}

\item{maxFBIteration}{integer}

\item{toleranceFB}{integer}

\item{useTranspose}{Boolean}
}
\value{
list of parameters to run the nGMCAs algorithm
}
\description{
specify the factorization algorithm parameters
}
