% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-function.R
\name{fun_get_rawdata}
\alias{fun_get_rawdata}
\title{Extract the isolation windows of SWATH data.}
\usage{
fun_get_rawdata(x)
}
\arguments{
\item{x}{MsExperiment object}
}
\value{
\code{data.frame} SWATH isolation windows.
}
\description{
Extract the isolation windows of SWATH data.
}
