% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2_dataGenerating.R
\name{generate_peaklist}
\alias{generate_peaklist}
\title{Build MS2 xics}
\usage{
generate_peaklist(
  msexp,
  mslevel = 1,
  ms2isowinL = TRUE,
  combineSpectra_arg,
  ppm = 6
)
}
\arguments{
\item{msexp}{\code{MsExperiment} object obtained from xcms or with \code{DIANMF::detect_xcms_peaks()}.}

\item{mslevel}{(optional) MSLevel to subset}

\item{ms2isowinL}{\code{logical} is \code{TRUE} to build the MS2 xics from raw data, else from xcms peaks.}

\item{combineSpectra_arg}{(optional) combineSpectra argument as a list}

\item{ppm}{\code{numeric} used for returning mzmin and mzmax values.}
}
\value{
\code{data.table} \code{data.frame} MS2 EICs.
}
\description{
Build MS2 xics
}
