% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exporting_fcts.R
\name{get_elutionprofile}
\alias{get_elutionprofile}
\title{Extract elution profile}
\usage{
get_elutionprofile(
  features.l = features,
  summary_dt = NULL,
  feature_id,
  sample_index = NULL,
  type = c("pure", "mixed")[1],
  max_method
)
}
\arguments{
\item{type}{\code{string} to retrieve the "pure" or "mixed" elution profiles.}
}
\value{
\code{data.table} of features with their optimal sources
}
\description{
Extract elution profile
}
