% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exporting_fcts.R
\name{get_feature_coord}
\alias{get_feature_coord}
\title{Export MS Spectra}
\usage{
get_feature_coord(
  features.l,
  summary_dt = NULL,
  feature_id,
  sample_index = NULL,
  max_method
)
}
\value{
\code{data.table} of features with their optimal sources
}
\description{
Export MS Spectra
}
