% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exporting_fcts.R
\name{get_feature_summary}
\alias{get_feature_summary}
\title{Export MS Spectra}
\usage{
get_feature_summary(features.l, max_method = c("contribution", "max_value")[1])
}
\arguments{
\item{features.l}{\code{list} obtainer with \code{DIANMF::DIANMF.f()}}

\item{max_method}{\code{string} to choose a method to select the optimal iteration:
\itemize{
\item \code{contribution}: select the first iteration with maximum contribution.
\item \code{max_value}: select the first iteration with maximum apex pure value.
}}
}
\value{
\code{data.table} of features with their optimal sources
}
\description{
Export MS Spectra
}
