% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-function.R
\name{get_isowin}
\alias{get_isowin}
\title{Get isolation windows}
\usage{
get_isowin(msexp)
}
\arguments{
\item{msexp}{MsExperiment object}
}
\value{
a numeric vector
}
\description{
Get isolation windows
}
