% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-function.R
\name{get_ms1_rtdiff}
\alias{get_ms1_rtdiff}
\title{Get rt diff between 2 MS1 scans}
\usage{
get_ms1_rtdiff(msexp)
}
\arguments{
\item{msexp}{MsExperiment object}
}
\value{
a data.table containing spectra infos
}
\description{
Get rt diff between 2 MS1 scans
}
