% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-function.R
\name{get_mslevels}
\alias{get_mslevels}
\title{Get unique msLevel from MsExperiment object}
\usage{
get_mslevels(msexp)
}
\arguments{
\item{msexp}{MsExperiment object}
}
\value{
an integer vector with available MSLevels
}
\description{
Get unique msLevel from MsExperiment object
}
