% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_matrices.R
\name{get_part2_matrix_X}
\alias{get_part2_matrix_X}
\title{extract the mixed MS2 data matrix, while trying to decrease the shifting between the MS1 and MS2 data}
\usage{
get_part2_matrix_X(
  pg.raw,
  fp.smp,
  idx.pg,
  scanTimes_mzRange,
  rt.info,
  fp.swath,
  int.filter,
  ppm.ms2.mtp,
  dia.feature
)
}
\arguments{
\item{pg.raw}{list raw peaks}

\item{fp.smp}{string file path}

\item{idx.pg}{numeric index of the MS1 peak}

\item{scanTimes_mzRange}{list MS1 & MS2 scans, rt of the scans and the mz range}

\item{rt.info}{list retention time information raw and extended for MS1 peak}

\item{fp.swath}{string SWATH file path}

\item{int.filter}{numeric minimum intensity accepted}

\item{ppm.ms2.mtp}{numeric used to extract the mz range}

\item{dia.feature}{list contains peaks information of the file

notes: when the peak may belon to scan index 1 I cant extend, so no retention time alignment
also some MS1 peaks wasnt fragmented in the MS2 steps so I cant extract there MS2 data; they doesnt exists}
}
\value{
list of MS2 information related the MS1 peak idx.pg; the MS2 matrix, rt and mz values of this peak
}
\description{
extract the mixed MS2 data matrix, while trying to decrease the shifting between the MS1 and MS2 data
}
