% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2_dataGenerating.R
\name{get_rawD_ntime}
\alias{get_rawD_ntime}
\title{Extract MS1 raw data}
\usage{
get_rawD_ntime(msexp, rt_range, sample_idx)
}
\arguments{
\item{msexp}{\code{MsExperiment} object obtained from xcms or with \code{DIANMF::detect_xcms_peaks()}.}

\item{rt_range}{retention time range.}

\item{sample_idx}{\code{numeric} sample index.}
}
\value{
\code{list} of \code{data.table} \code{data.frame} objects: raw_dt and time_dic.
}
\description{
Extract MS1 raw data
}
