% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-function.R
\name{get_spectra_index}
\alias{get_spectra_index}
\title{Get spectra index in MsExperiment}
\usage{
get_spectra_index(msexp)
}
\arguments{
\item{msexp}{MsExperiment object}
}
\value{
a data.table with scan infos
}
\description{
Get spectra index in MsExperiment
}
