% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-function.R
\name{get_spectra_values}
\alias{get_spectra_values}
\title{Get spectra value in MsExperiment}
\usage{
get_spectra_values(
  msexp,
  mslevel = NULL,
  isowin = NULL,
  combineSpectra_arg = NULL
)
}
\arguments{
\item{msexp}{\code{MsExperiment} object}

\item{mslevel}{(optional) MSLevel to subset}

\item{isowin}{(optional) isolation window to subset}

\item{combineSpectra_arg}{(optional) combineSpectra argument as a list}
}
\value{
a \code{data.table} containing spectra infos
}
\description{
Get spectra value in MsExperiment
}
