% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exporting_fcts.R
\name{get_spectrum_obj}
\alias{get_spectrum_obj}
\title{Extract spectrum (of class \code{Spectra}) from \code{data.table} \code{data.frame}}
\usage{
get_spectrum_obj(ms_Spectrum, feature_id)
}
\arguments{
\item{ms_Spectrum}{\code{data.table} \code{data.frame}}

\item{feature_id}{\code{character} feature id.}
}
\value{
\code{Spectra} object
}
\description{
Extract spectrum (of class \code{Spectra}) from \code{data.table} \code{data.frame}
}
