% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step3.2_NMF.R
\name{get_svd_first}
\alias{get_svd_first}
\title{Get first singular value using svd}
\usage{
get_svd_first(x, method = c("base", "fast", "svds"))
}
\arguments{
\item{x}{\code{matrix}.}

\item{method}{Method to compute svd: "base", "fast" or "svds".}
}
\value{
\code{ numeric(1)} relative error.
}
\description{
Get first singular value using svd
}
