% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_example.R
\name{get_test_peaks}
\alias{get_test_peaks}
\title{Perform peak picking on test data}
\usage{
get_test_peaks()
}
\value{
\code{XCMSnExp} object containing detected peaks from both mzML files
}
\description{
Perform peak picking on test data
}
