% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modified_Decometdia.R
\name{GetFeatures}
\alias{GetFeatures}
\title{get dia features from a mzML file}
\usage{
GetFeatures(
  obj,
  files,
  d.out,
  polarity = "positive",
  lc = "RPLC",
  ppm.pd = 6,
  sn = 0,
  peakwidth = c(5, 30),
  mzdiff = 3e-04,
  minfrac = 0.5,
  method.rector = "obiwarp",
  is.ms1.filter = FALSE,
  thr.ms1.filter = NULL,
  filed.ms1.filter = "into",
  is.plot.eic.feature = TRUE,
  is.check.eic = FALSE,
  nSlaves = 6,
  bw = 5,
  mzwid = 0.015
)
}
\arguments{
\item{obj}{new generated dia.Feature object}

\item{files}{list_of_strings mzML files paths}

\item{d.out}{path results path}

\item{polarity}{string 'POS' or 'NEG'}

\item{lc}{string Liquid chromatoghraphy column used for chromatograms separation}

\item{ppm.pd}{integer max mz devaition in consecutive scans}

\item{sn}{integer singl to noise ration}

\item{peakwidth}{integer range of the chromatogram peak width in seconds}

\item{mzdiff}{integer}

\item{minfrac}{integer}

\item{method.rector}{character}

\item{is.ms1.filter}{Boolean}

\item{thr.ms1.filter}{Boolean or NULL}

\item{filed.ms1.filter}{character}

\item{is.plot.eic.feature}{Boolean}

\item{is.check.eic}{Boolean}

\item{nSlaves}{integer}

\item{bw}{integer}

\item{mzwid}{integer}
}
\value{
dia.feature.RData and many other xcms objects
}
\description{
get dia features from a mzML file
}
