% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_metrics.R
\name{getReverseSearchingSimilarity}
\alias{getReverseSearchingSimilarity}
\title{Calculate reverse dot-product}
\usage{
getReverseSearchingSimilarity(measuredSpectra, librarySpectra, bin)
}
\arguments{
\item{measuredSpectra}{experimental spectrum.}

\item{librarySpectra}{reference spectrum.}

\item{bin}{\code{numeric} tolerance to merge.}
}
\value{
\code{numeric} reverse dot product score.
}
\description{
Calculate reverse dot-product
}
