% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_eics.R
\name{GetTIC}
\alias{GetTIC}
\title{get TIC}
\usage{
GetTIC(fn, rtcor = NULL)
}
\arguments{
\item{fn}{character files paths list}

\item{rtcor}{character c("raw", "corrected")}
}
\value{
TIC
}
\description{
get TIC
}
