% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_eics.R
\name{GetTICs}
\alias{GetTICs}
\title{get TICs}
\usage{
GetTICs(xsetc = NULL, fn.tic = "TICs.pdf", rt = c("raw", "corrected"))
}
\arguments{
\item{xsetc}{I should check}

\item{fn.tic}{character pdf name}

\item{rt}{character c("raw", "corrected")}
}
\value{
TICs
}
\description{
get TICs
}
