% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2_dataGenerating.R
\name{has_four_consecutive_non_zero}
\alias{has_four_consecutive_non_zero}
\title{Check if the vector has at least 4 consecutive non-zero values}
\usage{
has_four_consecutive_non_zero(row)
}
\arguments{
\item{row}{vector of \code{numeric} to evaluate.}
}
\value{
\code{Logical} \code{TRUE} if the vector contains at least 4  consecutive non-zero values, \code{FALSE} otherwise.
}
\description{
Check if the vector has at least 4 consecutive non-zero values
}
