% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2_dataGenerating.R
\name{has_n_consecutive_non_zero}
\alias{has_n_consecutive_non_zero}
\title{Check if the vector has at least 4 consecutive non-zero values}
\usage{
has_n_consecutive_non_zero(row, nscans = 4)
}
\arguments{
\item{row}{vector of \code{numeric} to evaluate.}

\item{nscans}{number of minimal consecutive value to check}
}
\value{
\code{Logical} \code{TRUE} if the vector contains at least n
consecutive non-zero values, \code{FALSE} otherwise.
}
\description{
Check if the vector has at least 4 consecutive non-zero values
}
