% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_mixed_data_as_matrices.R
\name{info_all}
\alias{info_all}
\title{extract rt, scans, eics for a specific MS1 peak}
\usage{
info_all(
  pg.raw,
  fp.smp,
  idx.pg,
  scanTimes_mzRange,
  rt.info,
  fp.swath,
  int.filter,
  ppm.ms2.mtp,
  dia.feature
)
}
\arguments{
\item{pg.raw}{list raw peaks}

\item{fp.smp}{file sample}

\item{idx.pg}{index of the peak}

\item{scanTimes_mzRange}{MS1 & MS2 scans information (scans times and spectrum for every scan)}

\item{rt.info}{rt information for every MS1 peak}

\item{fp.swath}{dictionary learning file}

\item{int.filter}{minimum intensity to filter}

\item{ppm.ms2.mtp}{ppm value to pick ms2 peaks}

\item{dia.feature}{dia features}
}
\value{
rt, scans indices, apex info of ms1 & ms2 peaks and eics raw, extended, and smooth
}
\description{
extract rt, scans, eics for a specific MS1 peak
}
