% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2.2_find_peaks_for_specific_precursors.R
\name{load_compounds_function.tb}
\alias{load_compounds_function.tb}
\title{load the compounds file as data-frames/table}
\usage{
load_compounds_function.tb(compound.c)
}
\arguments{
\item{compound.c}{file_path to load the spiked compounds table}
}
\value{
data-frame/table spiked compounds table
}
\description{
load the compounds file as data-frames/table
}
