% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step9.2_scores_for_matching.R
\name{match_pure_scores2}
\alias{match_pure_scores2}
\title{Match pure spectrum with some reference spectra to be identify.}
\usage{
match_pure_scores2(
  polarity,
  mz_prec,
  data_base,
  measured_spectra,
  mz_tol = NULL
)
}
\arguments{
\item{polarity}{\code{character} 'POS' or 'NEG'.}

\item{mz_prec}{\code{numeric(1)} precursor mz.}

\item{data_base}{reference/library database.}

\item{measured_spectra}{\code{data.frame} experimental pure spectrum.}

\item{mz_tol}{\code{numeric(1)} to bin the ions.}
}
\value{
\code{list} matching scores with some information of the matched library(reference) spectrum.
}
\description{
Match pure spectrum with some reference spectra to be identify.
}
