% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2_dataGenerating.R
\name{ms1Info}
\alias{ms1Info}
\title{Generate MS1 data and info}
\usage{
ms1Info(xic_dt_ms1)
}
\arguments{
\item{xic_dt_ms1}{\code{data.table} \code{data.frame} object obtained from xcms or with \code{DIANMF::build_ms1XICs()}.}
}
\value{
\code{list} of extracted MS1 data (MS1 matrix, deleted matrix, MS1 info).
}
\description{
Generate MS1 data and info
}
