% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_metrics.R
\name{ms2_matching}
\alias{ms2_matching}
\title{Match a measured spectrum with a database or a specific reference spectrum, by calculating the three scores: dot product, reverse dot product and fragment presence percentage.}
\usage{
ms2_matching(
  polarity,
  mz_prec,
  data_base,
  measured_spectra,
  mz_tol = NULL,
  idx = NA
)
}
\arguments{
\item{polarity}{\code{character} "POS" or "NEG".}

\item{mz_prec}{\code{numeric} precursor mz.}

\item{data_base}{spectra data base.}

\item{measured_spectra}{\code{data.frame} or \code{data.table} of two columns 'mz' and 'intensity'.}

\item{mz_tol}{\code{numeric} mz tolerance.}

\item{idx}{\code{numeric} index of a specific reference spectrum to match with.}
}
\value{
\code{data.table} matching results.
}
\description{
Match a measured spectrum with a database or a specific reference spectrum, by calculating the three scores: dot product, reverse dot product and fragment presence percentage.
}
