% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2_dataGenerating.R
\name{ms2Info}
\alias{ms2Info}
\title{Generate MS2 data and info}
\usage{
ms2Info(xic_dt_ms2)
}
\arguments{
\item{xic_dt_ms2}{\code{data.table} \code{data.frame} object obtained from xcms or with \code{DIANMF::build_ms2XICs()}.}
}
\value{
\code{list} of extracted MS2 data (MS2 matrix, MS2 info).
}
\description{
Generate MS2 data and info
}
