% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step4.1_initialization_methods.R
\name{neg}
\alias{neg}
\title{Negative projection.}
\usage{
neg(x)
}
\arguments{
\item{x}{vector of numeric.}
}
\value{
Projection of x in the negative orthant.
}
\description{
Negative projection.
}
