% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step4.3_nGMCAs_algo.R
\name{nGMCAs}
\alias{nGMCAs}
\title{apply nGMCAs on matrix Y to find A and S}
\usage{
nGMCAs(
  originalMatrix = X_final,
  rank,
  options = options.l,
  errors_print = FALSE
)
}
\arguments{
\item{originalMatrix}{description}

\item{rank}{numeric rank of factorization = number of pure compounds in the mixed data}

\item{options}{list of parameters}
}
\value{
data: list of 2 matrices A and S
}
\description{
apply nGMCAs on matrix Y to find A and S
}
