% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nndsvd.R
\name{nndsvd}
\alias{nndsvd}
\title{apply nndsvd on a matrix}
\usage{
nndsvd(A, k)
}
\arguments{
\item{A}{matrix}

\item{k}{integer: the rank}
}
\value{
list of two matrices H and W
}
\description{
apply nndsvd on a matrix
}
