% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step3.1_initialization_methods.R
\name{nndsvd_init}
\alias{nndsvd_init}
\title{nndsvd (non-negative double singular value decomposition).}
\usage{
nndsvd_init(X, rank)
}
\arguments{
\item{X}{\code{matrix}.}

\item{rank}{\code{numeric(1)} number of pure compounds in the mixed matrix; rank of factorization.}
}
\value{
\code{list} of two matrices H and W.
}
\description{
nndsvd (non-negative double singular value decomposition).
}
\examples{
 m <- matrix(c(1,0,0,0,1,1,1,1,1), nrow = 3, ncol = 3)
 r <- 2
 
nndsvd_init(X = m, rank = r)
}
