% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step3.1_initialization_methods.R
\name{norm_euc}
\alias{norm_euc}
\title{Calculates the square root of the sum of the squared elements in the input vector "x".}
\usage{
norm_euc(x)
}
\arguments{
\item{x}{vector of numeric.}
}
\value{
\code{numeric(1)}
}
\description{
Calculates the square root of the sum of the squared elements in the input vector "x".
}
