% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step5.1_prepare_pure_data.R
\name{plot_mixed_eics}
\alias{plot_mixed_eics}
\title{plot the mixed eics, experimental and real rt of the precursor}
\usage{
plot_mixed_eics(data.f1, rt.n_prec, rt.n_peak)
}
\arguments{
\item{data.f1}{data.frame}

\item{rt.n_prec}{numeric precursor apex real rt (the rt I have in the compounds table)}

\item{rt.n_peak}{numeric peak apex rt (peak apex rt detected by xcms)}
}
\value{
ggplot2 plot of mixed eics
}
\description{
plot the mixed eics, experimental and real rt of the precursor
}
