% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_metrics.R
\name{plot_spectra_vs}
\alias{plot_spectra_vs}
\title{Plot measured (mixed and pure) vs. library spectra of a feature}
\usage{
plot_spectra_vs(
  measuredSpectra_pure,
  measuredSpectra_mixed,
  librarySpectra,
  scores_pure,
  scores_mixed,
  mslevel
)
}
\arguments{
\item{measuredSpectra_pure}{\code{data.frame}}

\item{measuredSpectra_mixed}{\code{data.frame}}

\item{librarySpectra}{reference spectrum.}

\item{scores_pure}{\code{data.frame}}

\item{scores_mixed}{\code{data.frame}}

\item{mslevel}{"MS1" or "MS2}
}
\value{
ggplot2 plot
}
\description{
Plot measured (mixed and pure) vs. library spectra of a feature
}
