% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modified_Decometdia.R
\name{plotFeatureEIC}
\alias{plotFeatureEIC}
\title{plotFeatureEIC function}
\usage{
plotFeatureEIC(xset, is.check.eic = TRUE, extra = 30)
}
\arguments{
\item{xset}{xcms_object}

\item{is.check.eic}{Boolean}

\item{extra}{integer}
}
\value{
EIC plot
}
\description{
plotFeatureEIC function
}
