% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step3.1_initialization_methods.R
\name{pos}
\alias{pos}
\title{Positive projection.}
\usage{
pos(x)
}
\arguments{
\item{x}{vector of numeric.}
}
\value{
Projection of x in the positive orthant.
}
\description{
Positive projection.
}
