% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2.1_extract_eics.R
\name{PpmRange}
\alias{PpmRange}
\title{Create a mz range.}
\usage{
PpmRange(ref, ppm.n)
}
\arguments{
\item{ref}{\code{numeric} mz of the peak.}

\item{ppm.n}{\code{numeric(1)} defining the m/z tolerated deviation in parts per million (ppm).}
}
\value{
\code{numeric} mz range of a specific mz value.
}
\description{
Create a mz range.
}
