% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step5.1_prepare_pure_data.R
\name{prepare_mixed_eics}
\alias{prepare_mixed_eics}
\title{prepare mixed eics for plotting}
\usage{
prepare_mixed_eics(eics, retention_times)
}
\arguments{
\item{eics}{description}

\item{retention_times}{description}
}
\value{
data.frame
}
\description{
prepare mixed eics for plotting
}
