% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treating_data_after_factorization.R
\name{prepare_mixed_spectra}
\alias{prepare_mixed_spectra}
\title{prepare mixed spectra for plotting}
\usage{
prepare_mixed_spectra(spectra, mz_values)
}
\arguments{
\item{spectra}{description}

\item{mz_values}{description}
}
\value{
data.frame
}
\description{
prepare mixed spectra for plotting
}
