% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step1_dataPreparation.R
\name{prepare_mzMLfiles}
\alias{prepare_mzMLfiles}
\title{Load and prepare mzML files}
\usage{
prepare_mzMLfiles(input_dir)
}
\arguments{
\item{input_dir}{\code{character} mzML/mzXML files directory path.}
}
\value{
\code{data.table} \code{data.frame} contains mzML/mzXML files information.
}
\description{
Load and prepare mzML files
}
