% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treating_data_after_factorization.R
\name{prepare_pure_eics}
\alias{prepare_pure_eics}
\title{prepare pure eics for plotting}
\usage{
prepare_pure_eics(H, retention_times)
}
\arguments{
\item{H}{description}

\item{retention_times}{description}
}
\value{
data.frame
}
\description{
prepare pure eics for plotting
}
