% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step6.1_prepare_data.R
\name{prepare_pure_spectra}
\alias{prepare_pure_spectra}
\title{Prepare pure spectra for plotting.}
\usage{
prepare_pure_spectra(W)
}
\arguments{
\item{W}{Pure spectra \code{matrix}.}
}
\value{
Pure components spectra information \code{data.frame}.
}
\description{
Prepare pure spectra for plotting.
}
