% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step4.1_initialization_methods.R
\name{random_init}
\alias{random_init}
\title{Initialize matrices W and H randomly.}
\usage{
random_init(X, rank)
}
\arguments{
\item{X}{\code{matrix}.}

\item{rank}{\code{numeric(1)} number of pure compounds in the mixed matrix; rank of factorization.}
}
\value{
W and H random matrices.
}
\description{
Initialize matrices W and H randomly.
}
\examples{
 m <- matrix(c(1,0,0,0,1,1,1,1,1), nrow = 3, ncol = 3)
 r <- 2
 
random_init(X = m, rank = r)
}
