% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_general_info.R
\name{rt_info.f}
\alias{rt_info.f}
\title{extract rt information}
\usage{
rt_info.f(pg.raw, rt.extend, scanTimes_mzRange)
}
\arguments{
\item{pg.raw}{ms1 raw peaks}

\item{rt.extend}{rt value, to extend the peak rt range}

\item{scanTimes_mzRange}{ms1 & ms2 scans information (scans times and spectrum for every scan)}
}
\value{
rt information for every ms1 peak
}
\description{
extract rt information
}
