% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_general_info.R
\name{scan_time_info.f}
\alias{scan_time_info.f}
\title{extract ms1 & ms2 scans information}
\usage{
scan_time_info.f(fp.smp, dia.feature)
}
\arguments{
\item{fp.smp}{file sample}

\item{dia.feature}{dia features}
}
\value{
ms1 & ms2 scans information (scans times and spectrum for every scan)
}
\description{
extract ms1 & ms2 scans information
}
