% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2.2_find_peaks_for_specific_precursors.R
\name{search_for_ROIs_function.l}
\alias{search_for_ROIs_function.l}
\title{search for the peaks suggestions of every compound using the mz and retention time info}
\usage{
search_for_ROIs_function.l(compounds.l, matrix.n, rt_tol.n)
}
\arguments{
\item{compounds.l}{list empty, every level related to a compounds}

\item{matrix.n}{matrix of peaks}

\item{rt_tol.n}{numeric rt tolerance}

\item{mz_tol.n}{numeric mz tolerance}
}
\value{
filled list of peak suggestions for all levels
}
\description{
search for the peaks suggestions of every compound using the mz and retention time info
}
