% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step3.2_NMF.R
\name{soft_threshold.f}
\alias{soft_threshold.f}
\title{Apply soft-thresholding on a matrix.}
\usage{
soft_threshold.f(Y, threshold)
}
\arguments{
\item{Y}{\code{matrix}}

\item{threshold}{\code{numeric(1)} thresholding parameter.}
}
\value{
\code{matrix}
}
\description{
Apply soft-thresholding on a matrix.
}
