% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_database.R
\docType{data}
\name{sub_database}
\alias{sub_database}
\title{This is data to be included in my package}
\description{
This is data to be included in my package
}
\author{
Diana Karaki \email{zkaraki.diana@gmail.com}
}
\keyword{data}
