% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step3.1_initialization_methods.R
\name{subsample_init}
\alias{subsample_init}
\title{subSample initialization.}
\usage{
subsample_init(Y, rank, H_sub)
}
\arguments{
\item{Y}{\code{matrix}.}

\item{rank}{\code{numeric(1)} number of pure compounds in the mixed matrix; rank of factorization.}

\item{H_sub}{pure MS1 elution profiles \code{matrix}.}
}
\value{
A and S initializing matrices based on the data
}
\description{
subSample initialization.
}
\examples{
 m <- matrix(c(1,0,0,0,1,1,1,1,1), nrow = 3, ncol = 3)
 r <- 2

 H_sub <- matrix(c(0.9,0,0.8,1,0.06,0.1), nrow = r) 
 subsample_init(Y = t(m), rank = r, H_sub = H_sub) 
}
