% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step4_processing.R
\name{update_df}
\alias{update_df}
\title{Update df iteration columns}
\usage{
update_df(ms1_peaks, ms1_features, idx, iteration_number)
}
\arguments{
\item{ms1_peaks}{\code{data.frame} obtains from \code{DIANMF::extract_xcms_peaks()}}

\item{ms1_features}{\code{data.frame} obtains from \code{DIANMF::extract_xcms_features()}}

\item{idx}{\code{c(numeric)} row indexes}

\item{iteration_number}{\code{numeric(1)}}
}
\value{
A \code{list} containing updated \code{ms1_peaks} and \code{ms1_features} data.tables.
}
\description{
Update df iteration columns
}
