% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_algo.R
\name{updateS.f}
\alias{updateS.f}
\title{update the matrix S (it contains pure spectra)}
\usage{
updateS.f(Y, A, S_init, lambda, options)
}
\arguments{
\item{Y}{matrix}

\item{A}{matrix}

\item{S_init}{matrix}

\item{lambda}{numeric sparsity constraint}

\item{options}{list of some parameters}
}
\value{
new_S matrix updated pure spectra
}
\description{
update the matrix S (it contains pure spectra)
}
