% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_parameters.R
\name{xcms_parameters}
\alias{xcms_parameters}
\title{determine xcms parameters: for peak detection and alignment peaks from different samples}
\usage{
xcms_parameters(
  ppm = 6,
  sn = 0,
  peakwidth = c(5, 30),
  mzdiff = 3e-04,
  nSlaves = 6,
  rt.extend = 2.5,
  bw = 5,
  mzwid = 0.25,
  minsamp = 0,
  minfrac = 0.5
)
}
\arguments{
\item{ppm}{integer}

\item{sn}{integer}

\item{peakwidth}{integer}

\item{mzdiff}{integer}

\item{nSlaves}{integer}

\item{rt.extend}{integer

parameters to align peaks of multiple samples}

\item{bw}{integer}

\item{mzwid}{integer}

\item{minsamp}{integer}

\item{minfrac}{integer}
}
\value{
list of parameters for xcms
}
\description{
parameters to detect peaks
}
