% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2_dataGenerating.R
\name{build_XICs}
\alias{build_XICs}
\title{Build MS1 xics from peak list}
\usage{
build_XICs(peaks_dt, rawdt)
}
\arguments{
\item{peaks_i}{\code{data.table} \code{data.frame} targeted peaks in one rt_range.}

\item{raw_dt}{\code{data.frame} from \code{DIANMF::get_rawD_ntime}}
}
\value{
\code{data.table} \code{data.frame} MS1 EICs.
}
\description{
Build MS1 xics from peak list
}
