% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIANMF-package.R
\docType{package}
\name{DIANMF-package}
\alias{DIANMF}
\alias{DIANMF-package}
\title{DIANMF: Non-Negative Matrix Factorization To Identify SWATH DIA Metabolomic Data}
\description{
Identify metabolites from SWATH DIA MS2 metabolomic data by extracting their pure DIA MS2 spectra using Non-negative matrix Factorization. It is an end-to-end workflow called DIA-NMF that processes SWATH DIA files and identifies the detected compounds.
}
\keyword{internal}
