% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIANMF-package.R
\docType{package}
\name{DIANMF-package}
\alias{DIANMF}
\alias{DIANMF-package}
\title{DIANMF: Non-Negative Matrix Factorization To Identify SWATH DIA Metabolomic Data}
\description{
Sequential window acquisition of all theoretical data-independent acquisition mass spectra (SWATH DIA) is essential for information-rich spectral annotations in untargeted metabolomics. However, the acquired MS2 spectra are highly complex, posing significant annotation challenges. We have developed a DIANMF package that uses non-negative matrix factorization (NMF) to provide pure MS2 spectra that identify metabolomics in raw samples. DIANMF is based on NMF of the MS1 data before processing the MS2 data and on the assumption that for every metabolite, the existing information in the MS1 data correlates in the same sample with its MS2 data. Exploits this quantitative information to deconvolute complex SWATH DIA spectra.
}
\author{
\strong{Maintainer}: Diana Karaki \email{diana.karaki@cea.fr}

Authors:
\itemize{
  \item Sylvain Dechaumet \email{sylvain.dechaumet@cea.fr}
}

Other contributors:
\itemize{
  \item Etienne Thevenot \email{etienne.thevenot@cea.fr} [contributor]
}

}
\keyword{internal}
