% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIANMF-package.R
\docType{package}
\name{DIANMF-package}
\alias{DIANMF}
\alias{DIANMF-package}
\title{DIANMF: Non-Negative Matrix Factorization To Identify SWATH DIA Metabolomic Data}
\description{
This R package aims to extract pure MS1 and MS2 data raw mzML files.
}
\author{
\strong{Maintainer}: Diana Karaki \email{zkaraki.diana@gmail.com}

Authors:
\itemize{
  \item Sylvain Dechaumet \email{sylvain.dechaumet@cea.fr}
}

Other contributors:
\itemize{
  \item Etienne Thevenot \email{etienne.thevenot@cea.fr} [contributor]
}

}
\keyword{internal}
