% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step4_processing.R
\name{DIANMF.f}
\alias{DIANMF.f}
\title{Process xcms object}
\usage{
DIANMF.f(
  msexp,
  dir_out = FALSE,
  sample_idx = 1,
  MS2_ISOEACHL = TRUE,
  MS1MS2_L = FALSE,
  rank = 10,
  maximumIteration = 200,
  maxFBIteration = 100,
  toleranceFB = 1e-05,
  initialization_method = "nndsvd",
  errors_print = FALSE,
  method = "svds",
  rt_method = c("peak", "constant"),
  scan_rt_ext = 10,
  min_distance = 5,
  nscans = 4,
  featuresn = NULL,
  combineSpectra_arg = list(peaks = "intersect", ppm = 4, tolerance = 0.005, minProp =
    0.05),
  verbose = FALSE
)
}
\arguments{
\item{msexp}{\code{MsExperiment} object obtained from xcms or with \code{DIANMF::detect_xcms_peaks()}.}

\item{dir_out}{\code{logical} TRUE to save results, else FALSE}

\item{sample_idx}{NULL to process all files else, \code{numeric} index of the sample.}

\item{MS2_ISOEACHL}{\code{logical}}

\item{MS1MS2_L}{\code{logical}}

\item{rank}{\code{numeric(1)} number of pure compounds in the mixed matrix; rank of factorization.}

\item{maximumIteration}{\code{numeric(1)} maximum number of iterations.}

\item{maxFBIteration}{\code{numeric(1)} maximum number of iterations in the FB algorithm.}

\item{toleranceFB}{\code{numeric(1)} tolerance or stopping value of the FB algorithm.}

\item{initialization_method}{\code{character} to specify the initialization method: random, nndsvd or subSample.}

\item{errors_print}{\code{Logical} \code{TRUE} to print the error difference, \code{FALSE} otherwise.}

\item{method}{Method to compute svd: "base", "fast" or "svds"}

\item{rt_method}{c("peak", "constant")}

\item{scan_rt_ext}{\code{numeric}}

\item{min_distance}{\code{numeric}}

\item{nscans}{number of minimal consecutive value to check}

\item{featuresn}{\code{numeric} to limit the number if iteration (k) on each file for testing purpose, set it to NULL to extract all features.}

\item{combineSpectra_arg}{(optional) combineSpectra argument as a list}

\item{verbose}{\code{logical} to show execution messages}
}
\value{
\code{list}
}
\description{
Process xcms object
}
