% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step4_processing.R
\name{DIANMF.f}
\alias{DIANMF.f}
\title{Process xcms object}
\usage{
DIANMF.f(
  msexp,
  dir_out = FALSE,
  sample_idx = 1,
  MS2_ISOEACHL = TRUE,
  MS1MS2_L = TRUE,
  rank = 10,
  maximumIteration = 200,
  maxFBIteration = 100,
  toleranceFB = 1e-05,
  min_contrib = 0.6,
  initialization_method = "nndsvd",
  errors_print = FALSE,
  method = "svds",
  rt_method = c("peak", "constant"),
  scan_rt_ext = 10,
  min_distance = 5,
  nscans = 4,
  featuresn = NULL,
  clean_sources = TRUE,
  combineSpectra_arg = list(peaks = "intersect", ppm = 4, tolerance = 0.005, minProp =
    0.05),
  verbose = FALSE,
  BPPARAM = bpparam()
)
}
\arguments{
\item{msexp}{\code{MsExperiment} object obtained from xcms or with \code{DIANMF::detect_xcms_peaks()}.}

\item{dir_out}{\code{logical} TRUE to save results, else FALSE}

\item{sample_idx}{\code{NULL} to process all files else, \code{numeric} index of the sample.}

\item{MS2_ISOEACHL}{\code{logical}}

\item{MS1MS2_L}{\code{logical}}

\item{rank}{\code{numeric(1)} number of pure compounds in the mixed matrix; rank of factorization.}

\item{maximumIteration}{\code{numeric(1)} maximum number of iterations.}

\item{maxFBIteration}{\code{numeric(1)} maximum number of iterations in the FB algorithm.}

\item{toleranceFB}{\code{numeric(1)} tolerance or stopping value of the FB algorithm.}

\item{min_contrib}{\code{numeric} contribution proportion to keep ions (between 0 and 1).}

\item{initialization_method}{\code{character} to specify the initialization method: random, nndsvd or subSample.}

\item{errors_print}{\code{Logical} \code{TRUE} to print the error difference, \code{FALSE} otherwise.}

\item{method}{Method to compute svd: "base", "fast" or "svds".}

\item{rt_method}{c("peak", "constant")}

\item{scan_rt_ext}{\code{numeric} retention time extension of the rt window.}

\item{min_distance}{\code{numeric} mark the allowed distance from the rt window border for a peak to be entirely in the window or partially.}

\item{nscans}{number of minimal consecutive value to check}

\item{featuresn}{\code{numeric} to limit the number if iteration (k) on each file for testing purpose, set it to \code{NULL} to extract all features.}

\item{clean_sources}{\code{TRUE} to clean/remove sources with less than 0.6 contribution for any peaks, \code{FALSE} to keep all sources.}

\item{combineSpectra_arg}{(optional) combineSpectra argument as a list}

\item{verbose}{\code{logical} to show execution messages.}

\item{BPPARAM}{A \code{BiocParallelParam} object defining the parallelization backend to be used. Defaults to the value returned by \code{BiocParallel::bpparam()}.}
}
\value{
\code{list} of info of deconvoluted (pure) sources with the peaks-features data.
}
\description{
Process xcms object
}
