% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step3.2_NMF.R
\name{error_function}
\alias{error_function}
\title{Calculate the difference between Y an its approximation A*S.}
\usage{
error_function(Y, A, S)
}
\arguments{
\item{Y}{\code{matrix}}

\item{A}{\code{matrix} every column is an elution profile.}

\item{S}{\code{matrix} every row is a pure spectrum.}
}
\value{
\code{ numeric(1)} relative error.
}
\description{
Calculate the difference between Y an its approximation A*S.
}
