% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step4.3_nGMCAs_algo.R
\name{error_function}
\alias{error_function}
\title{calculate the difference between Y an its approximation A \%*\% S}
\usage{
error_function(Y, A, S)
}
\arguments{
\item{Y}{matrix}

\item{A}{matrix elution profiles}

\item{S}{matrix pure spectra}
}
\value{
relative error
}
\description{
calculate the difference between Y an its approximation A \%*\% S
}
