% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exporting_fcts.R
\name{export_featureSpect}
\alias{export_featureSpect}
\title{Extract \code{Spectra} object of a feature}
\usage{
export_featureSpect(
  features.l,
  feature_id,
  sample_index,
  type,
  method,
  max_method
)
}
\arguments{
\item{features.l}{\code{list} obtainer with \code{DIANMF::DIANMF.f()}}

\item{feature_id}{\code{character}}

\item{sample_index}{\code{numeric}}

\item{type}{\code{string} to retrieve the "pure" or "mixed" elution profiles.}

\item{method}{\code{string} to extract "all" sources or just the "best".}
}
\value{
\code{Spectra} object of a feature
}
\description{
Extract \code{Spectra} object of a feature
}
