% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exporting_fcts.R
\name{exportMSSpectra}
\alias{exportMSSpectra}
\title{Export MS pure spectra objects}
\usage{
exportMSSpectra(features.l)
}
\arguments{
\item{features.l}{\code{list} created by \code{DIANMF::DIANM.f}}
}
\value{
\code{Spectra} object of MS1 pure spectra.
}
\description{
Export MS pure spectra objects
}
