% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exporting_fcts.R
\name{exportMSSpectra}
\alias{exportMSSpectra}
\title{Extract MS Spectra of all identified features/peaks}
\usage{
exportMSSpectra(features.l, sample_index, type, method, max_method)
}
\arguments{
\item{features.l}{\code{list} obtained with \code{DIANMF::DIANMF.f()}}

\item{sample_index}{\code{numeric} sample index.}

\item{type}{\code{string} to retrieve the "pure" or "mixed" elution profiles.}

\item{method}{\code{string} to extract "all" sources or just the "best" source.}

\item{max_method}{\code{string} to choose a method to select the optimal iteration:
\itemize{
\item \code{contribution}: select the first iteration with maximum contribution.
\item \code{max_value}: select the first iteration with maximum apex pure value.
}}
}
\value{
\code{list} of Spectra objects for all identified features of one sample.
}
\description{
Extract MS Spectra of all identified features/peaks
}
