% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2.1_extract_eics.R
\name{extract_eics}
\alias{extract_eics}
\title{Extract the MS1 or MS2 eics from a list of spectra.}
\usage{
extract_eics(
  spectra_list,
  ppm.n = 7,
  apex_index,
  rt_index = TRUE,
  mz_range = TRUE
)
}
\arguments{
\item{spectra_list}{\code{list} of spectra.}

\item{ppm.n}{\code{numeric(1)} defining the m/z tolerated deviation in parts per million (ppm).}

\item{apex_index}{\code{integer} spectrum index which contains the peak apex.}

\item{rt_index}{\code{Logical} if \code{TRUE} use the real retention time axis of the peak, else \code{FALSE}.}

\item{mz_range}{\code{Logical} if \code{TRUE} and extracting MS1 data , if \code{TRUE} and extracting MS2 data , else \code{FALSE}.}
}
\value{
\code{matrix} contains extracted eic in every row.
}
\description{
Extract the MS1 or MS2 eics from a list of spectra.
}
