% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2.1_extract_eics.R
\name{extract_eics}
\alias{extract_eics}
\title{extract the MS1 or MS2 eics from a list of spectra related to a specific peak}
\usage{
extract_eics(
  spectra_list,
  ppm = 7,
  apex_index,
  rt_index = TRUE,
  mz_range = TRUE
)
}
\arguments{
\item{spectra_list}{}

\item{ppm}{}

\item{apex_index}{}

\item{rt_index}{}

\item{mz_range}{}
}
\value{
extracted eics
}
\description{
extract the MS1 or MS2 eics from a list of spectra related to a specific peak
}
