% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step1_dataPreparation.R
\name{extract_xcms_features}
\alias{extract_xcms_features}
\title{Extract MS1 features}
\usage{
extract_xcms_features(msexp, orderL_sample = NULL, quantifyL = FALSE)
}
\arguments{
\item{msexp}{\code{MsExperiment} object obtained from xcms or with \code{DIANMF::detect_xcms_peaks()}.}

\item{orderL_sample}{\code{character} sample name.}

\item{quantifyL}{Logical to perform quantification with \code{xcms::quantify()}}
}
\value{
\code{data.table} \code{data.frame} object.
}
\description{
Extract MS1 features
}
