% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exporting_fcts.R
\name{get_elutionprofile}
\alias{get_elutionprofile}
\title{Extract elution profile for a specific feature}
\usage{
get_elutionprofile(
  features.l = features,
  summary_dt = NULL,
  feature_id,
  sample_index = NULL,
  type = c("pure", "mixed")[1],
  method = c("all", "best")[2],
  max_method
)
}
\arguments{
\item{features.l}{\code{list} obtained with \code{DIANMF::DIANMF.f()}}

\item{summary_dt}{\code{data.table} of features with their optimal sources obtained from DIANMF::get_feature_summary().}

\item{feature_id}{\code{character} feature id.}

\item{sample_index}{\code{numeric(1)} sample index.}

\item{type}{\code{string} to retrieve the "pure" or "mixed" elution profiles.}

\item{method}{\code{string} to extract "all" sources or just the "best" source.}

\item{max_method}{\code{string} to choose a method to select the optimal iteration:
\itemize{
\item \code{contribution}: select the first iteration with maximum contribution.
\item \code{max_value}: select the first iteration with maximum apex pure value.
}}
}
\value{
\code{data.table} of features with their optimal sources
}
\description{
Extract elution profile for a specific feature
}
