% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exporting_fcts.R
\name{get_spectra}
\alias{get_spectra}
\title{Extract Spectra}
\usage{
get_spectra(
  features.l = features,
  summary_dt = NULL,
  feature_id = temp_ft[feature_i, featureid],
  sample_index = 1,
  type = c("pure", "mixed")[1],
  max_method
)
}
\arguments{
\item{type}{\code{string} to retrieve the "pure" or "mixed" elution profiles.}
}
\value{
\code{data.table} of features with their optimal sources
}
\description{
Extract Spectra
}
