% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-function.R
\name{get_spectra_values}
\alias{get_spectra_values}
\title{Get spectra value in MsExperiment}
\usage{
get_spectra_values(msexp)
}
\arguments{
\item{msexp}{MsExperiment object}
}
\value{
a data.table containing spectra infos
}
\description{
Get spectra value in MsExperiment
}
