% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_metrics.R
\name{GetPresenceSimilarity}
\alias{GetPresenceSimilarity}
\title{Calculate the matched fragments ratio}
\usage{
GetPresenceSimilarity(measuredSpectra, librarySpectra, bin)
}
\arguments{
\item{measuredSpectra}{experimental spectrum.}

\item{librarySpectra}{reference spectrum.}

\item{bin}{\code{numeric} tolerance to merge.}
}
\value{
\code{numeric} fragments ratio score.
}
\description{
Calculate the matched fragments ratio
}
