% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step9.2_scores_for_matching.R
\name{GetPresenceSimilarity}
\alias{GetPresenceSimilarity}
\title{Calculate the percentage of library spectrum fragments found in the experimental spectrum.}
\usage{
GetPresenceSimilarity(measuredSpectra, librarySpectra, bin)
}
\arguments{
\item{measuredSpectra}{\code{list} experimental spectrum.}

\item{librarySpectra}{\code{list} reference spectrum.}

\item{bin}{\code{numeric(1)} mz tolerance.}
}
\value{
\code{numeric(1)} percentage of library fragments in the experimental spectrum.
}
\description{
Calculate the percentage of library spectrum fragments found in the experimental spectrum.
}
