% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step5.2_scores_for_matching.R
\name{GetPresenceSimilarity}
\alias{GetPresenceSimilarity}
\title{presence of fragments score: calculate the percentage of library spectrum fragments found in the experimental spectrum}
\usage{
GetPresenceSimilarity(measuredSpectra, librarySpectra, bin)
}
\arguments{
\item{measuredSpectra}{list experimental spectrum}

\item{librarySpectra}{list reference spectrum}

\item{bin}{numeric mz tolerance}
}
\value{
numeric percentage of library fragments in the experimental spectrum
}
\description{
presence of fragments score: calculate the percentage of library spectrum fragments found in the experimental spectrum
}
