% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step5.2_scores_for_matching.R
\name{getReverseSearchingSimilarity}
\alias{getReverseSearchingSimilarity}
\title{inverse dot product score}
\usage{
getReverseSearchingSimilarity(measuredSpectra, librarySpectra, bin)
}
\arguments{
\item{measuredSpectra}{list experimental spectrum}

\item{librarySpectra}{list reference spectrum}

\item{bin}{numeric mz tolerance}
}
\value{
numeric inverse dot product
}
\description{
inverse dot product score
}
