% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step9.2_scores_for_matching.R
\name{getReverseSearchingSimilarity}
\alias{getReverseSearchingSimilarity}
\title{Inverse dot product score.}
\usage{
getReverseSearchingSimilarity(measuredSpectra, librarySpectra, bin)
}
\arguments{
\item{measuredSpectra}{\code{list} experimental spectrum.}

\item{librarySpectra}{\code{list} reference spectrum.}

\item{bin}{\code{numeric(1)} mz tolerance.}
}
\value{
\code{numeric(1)} inverse dot product.
}
\description{
Inverse dot product score.
}
\examples{
a <- data.frame( "mz_value" = seq(1:8), 'intensity' = c(0,400,1000,200,0,700,300,300) ) 
b <- data.frame( "mz_value" = seq(1:8), 'intensity' = c(100,0,1000,0,100,0,400,300))

getReverseSearchingSimilarity(measuredSpectra = a, librarySpectra = b, bin = 0.05)
}
