% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step7.2_scores_for_matching.R
\name{GetSimpleDotProductSimilarity}
\alias{GetSimpleDotProductSimilarity}
\title{dot product score}
\usage{
GetSimpleDotProductSimilarity(measuredSpectra, librarySpectra, bin)
}
\arguments{
\item{measuredSpectra}{list experimental spectrum}

\item{librarySpectra}{list reference spectrum}

\item{bin}{numeric mz tolerance}
}
\value{
numeric dot product
}
\description{
dot product score
}
