% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_metrics.R
\name{GetSimpleDotProductSimilarity}
\alias{GetSimpleDotProductSimilarity}
\title{Calculate the Dot product}
\usage{
GetSimpleDotProductSimilarity(measuredSpectra, librarySpectra, bin)
}
\arguments{
\item{measuredSpectra}{experimental spectrum.}

\item{librarySpectra}{reference spectrum.}

\item{bin}{\code{numeric} tolerance to merge.}
}
\value{
\code{numeric} dot product score.
}
\description{
Calculate the Dot product
}
