% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2.1_extract_eics.R
\name{isolationWindows.range}
\alias{isolationWindows.range}
\title{Extract the isolation windows of SWATH data.}
\usage{
isolationWindows.range(rawData.onDiskMSnExp)
}
\arguments{
\item{rawData.onDiskMSnExp}{\code{OnDiskMSnExp} object for \code{onDisk} mode.}
}
\value{
\code{data.frame} SWATH isolation windows.
}
\description{
Extract the isolation windows of SWATH data.
}
