% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2_extract_eics.R
\name{isolationWindows.range}
\alias{isolationWindows.range}
\title{Title}
\usage{
isolationWindows.range(rawData.onDiskMSnExp)
}
\arguments{
\item{rawData.onDiskMSnExp}{}
}
\value{
isolation windows
}
\description{
Title
}
