% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2.1_extract_eics.R
\name{isolationWindows.range}
\alias{isolationWindows.range}
\title{extract the isolation windows of SWATH data}
\usage{
isolationWindows.range(rawData.onDiskMSnExp)
}
\arguments{
\item{rawData.onDiskMSnExp}{}
}
\value{
isolation windows
}
\description{
extract the isolation windows of SWATH data
}
