% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step3.2_find_peaks_for_specific_precursors.R
\name{load_compounds_function.tb}
\alias{load_compounds_function.tb}
\title{Load the compounds table.}
\usage{
load_compounds_function.tb(compound.c)
}
\arguments{
\item{compound.c}{file path of the spiked compounds table}
}
\value{
data.frame/data.table spiked compounds table.
}
\description{
Load the compounds table.
}
\examples{
compounds_tb <- system.file("extdata", "Barbier_supplementary_tableS1.tsv", package = "DIANMF")

load_compounds_function.tb(compounds_tb)

}
