% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step5.2_scores_for_matching.R
\name{match_pure_scores2}
\alias{match_pure_scores2}
\title{match peaks with some reference spectra from Newdb.rds}
\usage{
match_pure_scores2(polarity, mz_precursor, data_base, measured_spectra)
}
\arguments{
\item{polarity}{string 'POS' or 'NEG'}

\item{mz_precursor}{numeric peak/MS1 precursor mz value}

\item{data_base}{reference/library database}

\item{measured_spectra}{list experimental pure spectrum}
}
\value{
list matching scores with some information of the matched library(reference) spectrum
}
\description{
match peaks with some reference spectra from Newdb.rds
}
