% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step4.1_initialization_methods.R
\name{neg}
\alias{neg}
\title{negative projection}
\usage{
neg(x)
}
\arguments{
\item{x}{vector of numerics}
}
\value{
vector the projection of x in the negative orthant
}
\description{
negative projection
}
