% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step4.3_nGMCAs_algo.R
\name{nGMCAs}
\alias{nGMCAs}
\title{apply nGMCAs on matrix Y to find A and S}
\usage{
nGMCAs(
  originalMatrix,
  rank,
  maximumIteration = 10,
  maxFBIteration = 10,
  toleranceFB = 1e-05,
  initialization_method = c("nndsvd", "random", "subSample"),
  H_sub = NULL,
  errors_print = FALSE
)
}
\arguments{
\item{originalMatrix}{description}

\item{rank}{numeric rank of factorization = number of pure compounds in the mixed data}

\item{maximumIteration}{numeric max number of iteration}

\item{maxFBIteration}{numeric FBS max number of iterations}

\item{toleranceFB}{numeric stopping value}

\item{initialization_method}{random, nndsvd or subSample}

\item{H_sub}{matrix elution profiles}

\item{errors_print}{Boolean}
}
\value{
data: list of 2 matrices A and S
}
\description{
apply nGMCAs on matrix Y to find A and S
}
