% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step3.2_NMF.R
\name{nGMCAs}
\alias{nGMCAs}
\title{Approximate Y by A and S.}
\usage{
nGMCAs(
  X.m,
  rank,
  maximumIteration = 50,
  maxFBIteration = 100,
  toleranceFB = 1e-05,
  initialization_method = c("nndsvd", "random", "subSample"),
  H_sub = NULL,
  errors_print = FALSE,
  scaleL = FALSE,
  method
)
}
\arguments{
\item{X.m}{\code{matrix} contains components' sources mixed up in an unknown but linear way. Every row is an elution profile from the candidate precursor and its fragment ions through the rt scans in the columns.}

\item{rank}{\code{numeric(1)} number of pure compounds in the mixed matrix; rank of factorization.}

\item{maximumIteration}{\code{numeric(1)} maximum number of iterations.}

\item{maxFBIteration}{\code{numeric(1)} maximum number of iterations in the FB algorithm.}

\item{toleranceFB}{\code{numeric(1)} tolerance or stopping value of the FB algorithm.}

\item{initialization_method}{\code{character} to specify the initialization method: random, nndsvd or subSample.}

\item{H_sub}{\code{matrix} of elution profiles used with subSample initialization method.}

\item{errors_print}{\code{Logical} \code{TRUE} to print the error difference, \code{FALSE} otherwise.}

\item{scaleL}{Logical to scale the data (row-wise) before applying NMF.}

\item{method}{Method to compute svd: "base", "fast" or "svds"}
}
\value{
\code{list} of 2 matrices A and S.
}
\description{
Approximate Y by A and S.
}
