% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nndsvd.R
\name{norm_euc}
\alias{norm_euc}
\title{calculates the square root of the sum of the squared elements in the input matrix "x"}
\usage{
norm_euc(x)
}
\arguments{
\item{x}{matrix}
}
\value{
numeric
}
\description{
calculates the square root of the sum of the squared elements in the input matrix "x"
}
