% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exporting_fcts.R
\name{plot_EluProfile}
\alias{plot_EluProfile}
\title{Export MS Spectra}
\usage{
plot_EluProfile(
  features.l,
  summary_dt = NULL,
  feature_id = NULL,
  sample_index = 1,
  log2L = FALSE,
  type = c("pure", "mixed")[1],
  max_method
)
}
\arguments{
\item{features.l}{\code{list} obtainer with \code{DIANMF::DIANMF.f()}}
}
\value{
\code{ggplot} of the mixed matrix for the asked feature.
}
\description{
Export MS Spectra
}
