% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exporting_fcts.R
\name{plot_feature}
\alias{plot_feature}
\title{Multiplot mixed/pure feature}
\usage{
plot_feature(
  features.l = features,
  summary_dt = NULL,
  feature_id = NULL,
  sample_index = 1,
  log2L = FALSE,
  max_method,
  method = c("all", "best")[2]
)
}
\arguments{
\item{features.l}{\code{list} obtainer with \code{DIANMF::DIANMF.f()}}

\item{log2L}{}
}
\value{
\code{ggplot} of the mixed spectra for the asked feature.
}
\description{
Multiplot mixed/pure feature
}
