% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_fcts.R
\name{plot_feature}
\alias{plot_feature}
\title{Multi-plot mixed/pure feature}
\usage{
plot_feature(
  features.l = features,
  summary_dt = NULL,
  feature_id = NULL,
  sample_index = 1,
  log2L = FALSE,
  max_method,
  method = c("all", "best")[2]
)
}
\arguments{
\item{features.l}{\code{list} obtained with \code{DIANMF::DIANMF.f()}}

\item{summary_dt}{\code{data.table} of features with their optimal sources obtained from DIANMF::get_feature_summary().}

\item{feature_id}{\code{character} feature id.}

\item{sample_index}{\code{numeric(1)} sample index.}

\item{log2L}{\code{logic} is \code{TRUE} for logarithmic transformation, else \code{FALSE}.}

\item{max_method}{\code{string} to choose a method to select the optimal iteration:
\itemize{
\item \code{contribution}: select the first iteration with maximum contribution.
\item \code{max_value}: select the first iteration with maximum apex pure value.
}}

\item{method}{\code{string} to extract "all" sources or just the "best" source.}
}
\value{
\code{ggplot} of the mixed spectra for the asked feature.
}
\description{
Multi-plot mixed/pure feature
}
