% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_fcts.R
\name{plot_rtWind_info}
\alias{plot_rtWind_info}
\title{Plot rt window with the factorization results}
\usage{
plot_rtWind_info(
  features.l,
  sample_index = 1,
  iteration_idx = 1,
  targets = NULL
)
}
\arguments{
\item{features.l}{\code{list} obtained with \code{DIANMF::DIANMF.f()}}

\item{sample_index}{\code{numeric} sample index.}

\item{iteration_idx}{\code{numeric} iteration index.}

\item{targets}{\code{data.frame} of two columns c(compound, mz) of co-eluting compounds, to highlight their eics in the mixed chromatograms plot.}
}
\value{
ggplot
}
\description{
Plot rt window with the factorization results
}
