% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exporting_fcts.R
\name{plot_Spectra}
\alias{plot_Spectra}
\title{Plot MS Spectra}
\usage{
plot_Spectra(
  features.l = features,
  summary_dt = NULL,
  feature_id = NULL,
  sample_index = 1,
  log2L = FALSE,
  type = c("pure", "mixed")[1],
  method = c("all", "best")[2],
  max_method
)
}
\arguments{
\item{features.l}{\code{list} obtainer with \code{DIANMF::DIANMF.f()}}

\item{log2L}{}
}
\value{
\code{ggplot} of the mixed spectra for the asked feature.
}
\description{
Plot MS Spectra
}
