% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exporting_fcts.R
\name{plot_Spectra}
\alias{plot_Spectra}
\title{Plot MS Spectra}
\usage{
plot_Spectra(
  features.l = features.l,
  summary_dt = NULL,
  feature_id = NULL,
  sample_index = 1,
  log2L = FALSE,
  type = c("pure", "mixed")[1],
  method = c("all", "best")[2],
  max_method
)
}
\arguments{
\item{features.l}{\code{list} obtained with \code{DIANMF::DIANMF.f()}}

\item{summary_dt}{\code{data.table} of features with their optimal sources obtained from DIANMF::get_feature_summary().}

\item{feature_id}{\code{character} feature id.}

\item{sample_index}{\code{numeric} sample index.}

\item{log2L}{\code{logic} is \code{TRUE} for logarithmic transformation, else \code{FALSE}.}

\item{type}{\code{string} to retrieve the "pure" or "mixed" elution profiles.}

\item{method}{\code{string} to extract "all" sources or just the "best" source.}

\item{max_method}{\code{string} to choose a method to select the optimal iteration:
\itemize{
\item \code{contribution}: select the first iteration with maximum contribution.
\item \code{max_value}: select the first iteration with maximum apex pure value.
}}
}
\value{
\code{ggplot} of the mixed spectra for the asked feature.
}
\description{
Plot MS Spectra
}
