% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step4.1_initialize.R
\name{pos}
\alias{pos}
\title{positive projection}
\usage{
pos(x)
}
\arguments{
\item{x}{vector of numeric}
}
\value{
vector the projection of x in the positive orthant
}
\description{
positive projection
}
