% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2.1_extract_eics.R
\name{PpmRange}
\alias{PpmRange}
\title{Create a mz range}
\usage{
PpmRange(ref, ppm)
}
\arguments{
\item{ref}{numeric mz of the peak/precursor}

\item{ppm}{numeric}
}
\value{
mz range of a specific mz value
}
\description{
Create a mz range
}
