% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2_extract_eics.R
\name{PpmRange}
\alias{PpmRange}
\title{create a mz range}
\usage{
PpmRange(ref, ppm)
}
\arguments{
\item{ref}{}

\item{ppm}{}
}
\value{
mz range of a specific mz value
}
\description{
create a mz range
}
