% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2_dataGenerating.R
\name{PpmRange}
\alias{PpmRange}
\title{Create a mz range}
\usage{
PpmRange(ref, ppm.n)
}
\arguments{
\item{ref}{\code{numeric} mz of the peak.}

\item{ppm.n}{\code{numeric} defining the m/z tolerated deviation in parts per million (ppm).}
}
\value{
\code{numeric} mz range.
}
\description{
Create a mz range
}
