% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step6.1_prepare_data.R
\name{prepare_pure_eics}
\alias{prepare_pure_eics}
\title{Prepare pure elution profiles (chromatograms) for plotting.}
\usage{
prepare_pure_eics(H, rt)
}
\arguments{
\item{H}{Elution profiles \code{matrix}.}

\item{rt}{\code{numeric} vector chromatograms retention time values in H.}
}
\value{
Pure components chromatograms information \code{data.frame}.
}
\description{
Prepare pure elution profiles (chromatograms) for plotting.
}
