% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step6.1_prepare_and_plot_data.R
\name{prepare_pure_eics}
\alias{prepare_pure_eics}
\title{prepare pure elution profiles for plotting}
\usage{
prepare_pure_eics(H, rt)
}
\arguments{
\item{H}{matrix contains elution profiles}

\item{rt}{numeric vector contains the fragments retention time values of the eics in H}
}
\value{
data.frame
}
\description{
prepare pure elution profiles for plotting
}
