% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step5.1_prepare_pure_data.R
\name{prepare_pure_spectra}
\alias{prepare_pure_spectra}
\title{prepare pure spectra for plotting}
\usage{
prepare_pure_spectra(W, mz_values)
}
\arguments{
\item{W}{description}

\item{mz_values}{description}
}
\value{
data.frame
}
\description{
prepare pure spectra for plotting
}
