% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step6.1_prepare_and_plot_data.R
\name{prepare_pure_spectra}
\alias{prepare_pure_spectra}
\title{prepare pure spectra for plotting}
\usage{
prepare_pure_spectra(W)
}
\arguments{
\item{W}{description}
}
\value{
data.frame
}
\description{
prepare pure spectra for plotting
}
