% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step4.1_initialize.R
\name{random_init}
\alias{random_init}
\title{Initialize matrices W and H randomly.}
\usage{
random_init(A, rank)
}
\arguments{
\item{A}{}

\item{rank}{}
}
\value{
W and H randon matrices
}
\description{
Initialize matrices W and H randomly.
}
