% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step4.1_initialization_methods.R
\name{random_init}
\alias{random_init}
\title{Initialize matrices W and H randomly.}
\usage{
random_init(X, rank)
}
\arguments{
\item{X}{}

\item{rank}{}
}
\value{
W and H random matrices
}
\description{
Initialize matrices W and H randomly.
}
