% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-function.R
\name{search_features}
\alias{search_features}
\title{Search for features related to specific compounds}
\usage{
search_features(feature_dt, dt, rttol = 5, ppm = 5)
}
\arguments{
\item{feature_dt}{feature summary}

\item{dt}{database to match (mandatory rt and mz columns)}

\item{rttol}{rt tolerance (in seconds)}

\item{ppm}{ppm tolerance for m/Z search}
}
\value{
a data.table
}
\description{
Search for features related to specific compounds
}
