% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step3.2_find_peaks_for_specific_precursors.R
\name{search_for_ROIs_function.l}
\alias{search_for_ROIs_function.l}
\title{Search for the peaks suggestions for compound using the mz and retention time info.}
\usage{
search_for_ROIs_function.l(
  compounds.df,
  ms1_peaks,
  mz_tol.n = 0.05,
  rt_tol.n = 5
)
}
\arguments{
\item{compounds.df}{\code{data.frame}}

\item{ms1_peaks}{\code{matrix} or \code{data.frame} contains MS1 peaks to be identified. It should contains these columns: mz, mzmin, mzmax, rt, rtmin, rtmax, into.}

\item{mz_tol.n}{\code{numeric(1)} mz tolerance.}

\item{rt_tol.n}{\code{numeric(1)} retention time tolerance.}
}
\value{
MS1 selected peaks \code{matrix}.
}
\description{
Search for the peaks suggestions for compound using the mz and retention time info.
}
