% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_algo.R
\name{soft_threshold.f}
\alias{soft_threshold.f}
\title{apply soft-thresholding on a matrix}
\usage{
soft_threshold.f(X, threshold)
}
\arguments{
\item{X}{matrix}

\item{threshold}{numeric thresholding parameter}
}
\value{
matrix
}
\description{
apply soft-thresholding on a matrix
}
