% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_algo.R
\name{updateA.f}
\alias{updateA.f}
\title{update the matrix A (it contains elution profiles)}
\usage{
updateA.f(Y, A_init, S, lambda, options)
}
\arguments{
\item{Y}{matrix}

\item{A_init}{matrix}

\item{S}{matrix}

\item{lambda}{numeric sparsity constraint}

\item{options}{list of some parameters}
}
\value{
new_A matrix updated elution profiles
}
\description{
update the matrix A (it contains elution profiles)
}
