% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step4.3_nGMCAs_algo.R
\name{updateA.f}
\alias{updateA.f}
\title{update the matrix A (it contains elution profiles)}
\usage{
updateA.f(Y, A_init, S, maxFBIteration, toleranceFB)
}
\arguments{
\item{Y}{matrix}

\item{A_init}{matrix}

\item{S}{matrix}

\item{maxFBIteration}{numeric max number of iteration}

\item{toleranceFB}{numeric stopping value}
}
\value{
new_A matrix updated elution profiles
}
\description{
update the matrix A (it contains elution profiles)
}
