% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step3.2_NMF.R
\name{updateA.f}
\alias{updateA.f}
\title{Update A}
\usage{
updateA.f(
  Y,
  A_init,
  S,
  maxFBIteration,
  toleranceFB,
  method = c("base", "fast")
)
}
\arguments{
\item{Y}{\code{matrix}.}

\item{A_init}{\code{matrix} every column is an elution profile.}

\item{S}{\code{matrix} every row is a pure spectrum.}

\item{maxFBIteration}{\code{numeric(1)} maximum number of iterations in the FB algorithm.}

\item{toleranceFB}{\code{numeric(1)} tolerance or stopping value of the FB algorithm.}

\item{method}{Method to compute svd: "base", "fast" or "svds".}
}
\value{
\code{matrix} A of updated elution profiles.
}
\description{
Update A
}
