% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step4.3_nGMCAs_algo.R
\name{updateS.f}
\alias{updateS.f}
\title{update the matrix S (it contains pure spectra)}
\usage{
updateS.f(Y, A, S_init, lambda, maxFBIteration, toleranceFB)
}
\arguments{
\item{Y}{matrix}

\item{A}{matrix elution profiles}

\item{S_init}{matrix pure spectra}

\item{lambda}{numeric sparsity parameter of the pure spectra}

\item{maxFBIteration}{numeric FBS max number of iteration}

\item{toleranceFB}{numeric stopping value}
}
\value{
S
}
\description{
update the matrix S (it contains pure spectra)
}
