% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step4.3_nGMCAs_algo.R
\name{updateS.f}
\alias{updateS.f}
\title{Update S.}
\usage{
updateS.f(
  Y,
  A,
  S_init,
  lambda,
  maxFBIteration,
  toleranceFB,
  method = c("base", "fast")[2]
)
}
\arguments{
\item{Y}{\code{matrix}}

\item{A}{\code{matrix} every column is an elution profile.}

\item{S_init}{\code{matrix} every row is a pure spectrum.}

\item{lambda}{\code{numeric(1)} sparsity parameter of the pure spectra.}

\item{maxFBIteration}{\code{numeric(1)} maximum number of iterations in the FB algorithm.}

\item{toleranceFB}{\code{numeric(1)} tolerance or stopping value of the FB algorithm.}

\item{method}{Method to compute svd: "base", "fast" or "svds"}
}
\value{
\code{matrix} S of updated spectra.
}
\description{
Update S.
}
