% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step4.3_nGMCAs_algo.R
\name{updateS.f}
\alias{updateS.f}
\title{update the matrix S (it contains pure spectra)}
\usage{
updateS.f(Y, A, S_init, lambda, maxFBIteration, toleranceFB)
}
\arguments{
\item{Y}{}

\item{A}{}

\item{S_init}{}

\item{lambda}{}

\item{maxFBIteration}{}

\item{toleranceFB}{}
}
\value{
S
}
\description{
update the matrix S (it contains pure spectra)
}
